fix: ACF wire-format precision — ReadSizeOrSegment width/naming, op-bit selection, acf_msg_length quadlet semantics
Three related acf.rs wire-format defects, fixed together since all three
touch ByteMessageInfo's shared byte_message_info layout:

- RELAY-TERM-02: renames acf::ReadSizeOrSegmentNum/
  ByteMessageInfo::read_size_segment_num to the RELAY specification's
  canonical §15.5 name (ReadSizeOrSegment/read_size_segment) and widens the
  field from an 8-bit placeholder to the spec's 16-bit width. The extra
  byte comes from what byte_message_info's encoding previously left as a
  reserved zero byte, so the 8-byte header length is unchanged. This
  rename/widen ripples into every module that reused the type for its own
  local read-size concept (uart's RX read-size config, the C FFI mirror in
  capi.rs, fragment.rs's segment_num tracking, and the Endpoint trait's
  own read_size parameter across every decorator) -- all updated to u16.

- rust-RCP-05: ByteMessageInfo::read_size()/segment_num() now select the
  ReadSizeOrSegment field's interpretation by the already-decoded op flag
  (read vs. write), returning None on the side op does not indicate,
  instead of two same-bit accessor views that returned the same value
  regardless of which interpretation actually applied.
  ReadSizeOrSegment::as_read_size/as_segment_num remain as unconditional
  raw-value views for contexts with no live op bit to gate on (e.g.
  uart's own local read_size field).

- rust-RCP-N2-05: encode_acf_abb/encode_acf_gbb now derive acf_msg_length
  from the real payload length (ceil(payload_len/4) quadlets) rather than
  trusting whatever the caller's ByteMessageInfo carried, rejecting
  payloads too large for the 11-bit field. decode_acf_abb/decode_acf_gbb
  now cross-check the decoded acf_msg_length against the actual payload
  bytes present and reject a mismatch, rather than leaving the field
  unvalidated. e2e.rs's CRC-32 coverage-buffer construction -- which
  deliberately pre-adjusts acf_msg_length by one quadlet for a value that
  is never itself transmitted -- now assembles its scratch buffer via the
  lower-level encode_byte_message_info directly rather than through the
  now-deriving encode_acf_abb/encode_acf_gbb, since that pre-adjustment
  has nothing to do with a real wire frame's actual length.

Closes #98, closes #102, closes #117

+++ src/deadline.rs
@@ -85,7 +85,7 @@ impl DeadlineEndpoint {
     // fusa:req REQ-DL-005
     pub fn read_with_deadline(
         &self,
-        read_size: u8,
+        read_size: u16,
         timeout: Option<Duration>,
     ) -> Result<Vec<u8>, RcpError> {
         let _effective = self.effective(timeout)?;
@@ -112,7 +112,7 @@ impl Endpoint for DeadlineEndpoint {
     }
 
     // fusa:req REQ-DL-006
-    fn read(&self, read_size: u8) -> Result<Vec<u8>, RcpError> {
+    fn read(&self, read_size: u16) -> Result<Vec<u8>, RcpError> {
         self.inner.read(read_size)
     }
